// src/app/manage/vehicle-manage/model/upkeep-record-res.model.ts
export class UpkeepRecordRes {
    id: number = null;
    acceptancePerson: string = null;
    chargePerson: string = null;
    planUpkeepTime: Date = null;
    plateNumber: string = null;
    upkeepCompany: string = null;
    upkeepContent: string = null;
    upkeepCost: number = null;
    upkeepMileage: number = null;
    upkeepPerson: string = null;
    upkeepTime: Date = null;
    upkeepTimeCost: number = null;
    vehicleId: number = null;
}